requirements_all.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
auraloss==0.4.0
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docker-pycreds==0.4.0
docstring_parser==0.16
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.1
hypy-utils==1.0.22
idna==3.6
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonargparse==4.27.7
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.1
lightning-utilities==0.11.2
llvmlite==0.42.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
neutone_sdk @ git+https://github.com/Neutone/neutone_sdk.git@c69108d6ad11910afdc3ce4ead0a647456e2addb
numba==0.59.1
numpy==1.23.5
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pillow==10.3.0
pip-chill==1.0.3
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
resampy==0.4.3
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.11.2
sentry-sdk==1.44.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchmetrics==1.3.2
torchvision==0.16.2
tqdm==4.66.2
typeshed_client==2.5.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.6
Werkzeug==3.0.2
yarl==1.9.4
